Restructure mdx-components
