+++ apps/site/src/components/mdx/modules/graph/graph-module-message-list.tsx
@@ -1,7 +1,7 @@
 import { ModuleDefinition } from "@blockprotocol/core";
 import { graphModuleJson } from "@blockprotocol/graph/graph-module-json";
 
-import { ModuleMessageList } from "../module-message-list";
+import { ModuleMessageList } from "../../../module-message-list";
 
 // @todo generalize this - need a way of having the module JSON at build time without importing like this,
 // e.g. pass a URL from the MDX component which is fetched at build time
